+++ src/main.rs
@@ -2,7 +2,6 @@
 #[rustfmt::skip] mod domain;
 #[rustfmt::skip] mod cli; // 0
 #[rustfmt::skip] mod database; // 1
-#[rustfmt::skip] mod query; // 5
 #[rustfmt::skip] mod http; // 6
 #[rustfmt::skip] mod presets; // 7